+++ MyHebrewBible/MyHebrewBible.Client/Components/Report/ReportModel.cs
@@ -2,15 +2,15 @@
 
 public class ReportModel
 {
-	public long ID { get; init; }
+	public int ID { get; init; }
 	public string? BCV { get; init; }
-	public long Verse { get; init; }
+	public int Verse { get; init; }
 	public string? VerseOffset { get; init; }
 	public string? KJV { get; init; }
 	public string? DescH { get; init; }
 	public string? DescD { get; init; }
-	public long BookID { get; init; }
-	public long Chapter { get; init; }
+	public int BookID { get; init; }
+	public int Chapter { get; init; }
 	public List<WordPart>? WordPartList { get; init; }
 }
 
@@ -18,9 +18,9 @@ public record WordPart
 {
 	public int Id { get; init; }                  
 	public string? BCV { get; init; }
-	public long BookID { get; init; }             
-	public long Chapter { get; init; }            
-	public long Verse { get; init; }              
+	public int BookID { get; init; }             
+	public int Chapter { get; init; }            
+	public int Verse { get; init; }              
 	public int ScriptureID { get; init; }         
 	public int WordCount { get; init; }
 	public int SegmentCount { get; init; }
