+++ MyHebrewBible/MyHebrewBible.Client/Features/Articles/Search/HebrewStrongsFormVM.cs
@@ -5,5 +5,5 @@ namespace MyHebrewBible.Client.Features.Articles.Search;
   public class HebrewStrongsFormVM
   {
 	[Range(1, 8674, ErrorMessage = "Hebrew Strong's must be between 1 and 8674.")]
-	public long Number { get; set; } // = 1;
+	public int Number { get; set; } // = 1;
 }
