+++ MyHebrewBible/MyHebrewBible.Client/Features/Articles/Search/SearchType.cs
@@ -2,4 +2,4 @@
 
 namespace MyHebrewBible.Client.Features.Articles;
 
-public record SearchType(StrongsType StrongsType, long StrongsNumber);
+public record SearchType(StrongsType StrongsType, int StrongsNumber);
