+++ MyHebrewBible/MyHebrewBible.Client/Features/Parasha/TableOfContents/ScriptureSelected.cs
@@ -1,3 +1,3 @@
 ﻿namespace MyHebrewBible.Client.Features.Parasha.TableOfContents;
 
-public record ScriptureSelected(long ScriptureId, Enums.ShowSection ShowSection );
+public record ScriptureSelected(int ScriptureId, Enums.ShowSection ShowSection );
