+++ MyHebrewBible/MyHebrewBible.Client/Features/Parasha/TableOfContents/TOC_VM.cs
@@ -6,7 +6,7 @@ public class TOC_VM()
 {
 	public List<TocRecord>? TableOfContentList { get; set; }
 
-	public long GetTorahScriptureId()
+	public int GetTorahScriptureId()
 	{
 		return TableOfContentList!.Where(w => w.SectionId == ShowSection.Torah.Value).Select(s => s.ScriptureID_Beg).SingleOrDefault(1);
 	}
