+++ MyHebrewBible/MyHebrewBible.Client/Features/VerseList/Section/ScriptureIdAndVerseNumber.cs
@@ -1,3 +1,3 @@
 ﻿namespace MyHebrewBible.Client.Features.VerseList.Section;
 
-public record ScriptureIdAndVerseNumber(long ScriptureID, long VerseNumber);
+public record ScriptureIdAndVerseNumber(int ScriptureID, int VerseNumber);
