+++ Utility/VerseRange.cs
@@ -1,4 +1,4 @@
 ﻿namespace Utility;
 
-public record VerseRange(BibleBook BibleBook, string ChapterVerse, long	BegId, long EndId);
+public record VerseRange(BibleBook BibleBook, string ChapterVerse, int BegId, int EndId);
